Allow callbacks to the invoking call
So far it was possible to run RpcConnection#call to wait for RPC calls from the other side.
In addition each RPC call could define a block to receive RPC calls.
However it was not defined which block received which RPC call.
There was no way to call back to a specific block.

This is now possible by Call#call_back.