Better compatibility on Windows
IO.popen converts \n to \r\n, even when called with binmode enabled.
To make communication safe on Windows text mode IOs we ignore the \r on line end now.
And we escape \r in data fields in addition to the other special characters.

Also revert to opening TCPSocket connection with localhost, as connecting to "::" or "0.0.0.0" fails on Windows.