Change the anonymous callback to return a Promise object as well
Previously the conn.call with only a block was excluded to return a Promise object.
This was because it doesn't send an RPC call.
But it makes sense to return a Promise, since it ensures that the anonymous callback is registered, when the Promise is returned.
Otherwise when the anonymous callback is started in a thread, it could lead to a raise like in this test case.

  1) Error:
TestRpcConnection#test_anonymous_callback:
Ccrpc::RpcConnection::NoCallbackDefined: call to "callbackaa" was received, but there is no Ccrpc::RpcConnection#call running

This breaks backward compatibility for anonymous callback registration with enabled lazy_answers.
It now requires to call `.itself` on the returned promise object to start waiting for callbacks like so:

```ruby
    c = Ccrpc::RpcConnection.new(r, w, lazy_answers: true)
    c.call do |call|
      # call processing
    end.itself
```